+++ delivery/src/main/java/com/berry/delivery/presentation/controller/DeliveryController.java
@@ -2,7 +2,7 @@ package com.berry.delivery.presentation.controller;
 
 import com.berry.common.response.ApiResponse;
 import com.berry.common.response.ResSuccessCode;
-import com.berry.delivery.application.service.DeliveryService;
+import com.berry.delivery.application.service.delivery.DeliveryService;
 import com.berry.delivery.domain.model.DeliveryStatus;
 import com.berry.delivery.presentation.dto.DeliveryDto;
 import com.berry.delivery.presentation.dto.request.DeliveryCreateRequest;
